#Replaced current tests that were responsible for comment reply from the "CommentCreation" to the new test class "EcoNewsCommentReplyTests"
