+++ src/test/java/com/softserve/edu/greencity/ui/tests/comments/CommentCreation.java
@@ -6,7 +6,6 @@ import com.softserve.edu.greencity.data.econews.NewsData;
 import com.softserve.edu.greencity.data.econews.NewsDataRepository;
 import com.softserve.edu.greencity.ui.pages.common.CommentComponent;
 import com.softserve.edu.greencity.ui.pages.common.CommentPart;
-import com.softserve.edu.greencity.ui.pages.common.ReplyComponent;
 import com.softserve.edu.greencity.ui.pages.econews.EcoNewsPage;
 import com.softserve.edu.greencity.ui.tests.runner.GreenCityTestRunner;
 import com.softserve.edu.greencity.ui.tools.jdbc.services.EcoNewsService;
@@ -38,8 +37,7 @@ public class CommentCreation extends GreenCityTestRunner {
                 .switchToSingleNewsPageByParameters(news)
                 .getCommentPart()
                 .addComment("First Comment")
-                .chooseCommentByNumber(0)
-                .addReply("First reply");
+                .chooseCommentByNumber(0);
         signOutByStorage();
     }
 
@@ -48,24 +46,6 @@ public class CommentCreation extends GreenCityTestRunner {
         EcoNewsPage ecoNewsPage = loadApplication().navigateMenuEcoNews();
         getEcoNewsService().deleteNewsByTitle(news.getTitle());
         softAssert.assertFalse(ecoNewsPage.refreshPage().isNewsDisplayedByTitle(news.getTitle()));
-    }
-
-    @Test
-    @Description("GC-958")
-    public void loggedUserCanPublishReply() {
-        logger.info("Verify that logged user can publish reply starts");
-
-        String replyText = "Test reply";
-        ReplyComponent replyComponent = loadApplication()
-                .loginIn(getTemporaryUser())
-                .navigateMenuEcoNews()
-                .switchToSingleNewsPageByParameters(news)
-                .getCommentPart()
-                .chooseCommentByNumber(0)
-                .addReply(replyText).openReply().chooseReplyByNumber(0);
-
-        softAssert.assertEquals(replyText, replyComponent.getReplyComment().getText());
-        softAssert.assertAll();
     }
 
     @Test
@@ -138,7 +118,6 @@ public class CommentCreation extends GreenCityTestRunner {
                 .addComment(commentText)
                 .chooseCommentByNumber(0)
                 .clickDeleteCommentButton();
-
         String lastCommentText = loadApplication()
                 .navigateMenuEcoNews()
                 .switchToSingleNewsPageByParameters(news)
